backend/requirements.txt
flask
flask_cors
google-cloud-storage
werkzeug
psycopg2-binary
